TerminalScreen: fix append multiline, wrong splitting of lines...
Detected only on the terminal title, it was beeing split incorrectly, but i couldn't find out why.
The logic would check the buffer for a newline char before appending a new string that started with a newline char. if the buffer had one, a new buffer would be created, else it would be appended on the current buffer. Anyway I think its best to drop this logic alltogether and always create a new buffer if the text-to-append starts with the newline char.